+++ services/network/mgmt/2020-11-01/network/CHANGELOG.md
@@ -16,6 +16,18 @@
 
 ### New Constants
 
+1. InterfaceMigrationPhase.InterfaceMigrationPhaseAbort
+1. InterfaceMigrationPhase.InterfaceMigrationPhaseCommit
+1. InterfaceMigrationPhase.InterfaceMigrationPhaseCommitted
+1. InterfaceMigrationPhase.InterfaceMigrationPhaseNone
+1. InterfaceMigrationPhase.InterfaceMigrationPhasePrepare
+1. InterfaceNicType.InterfaceNicTypeElastic
+1. InterfaceNicType.InterfaceNicTypeStandard
+1. PublicIPAddressMigrationPhase.PublicIPAddressMigrationPhaseAbort
+1. PublicIPAddressMigrationPhase.PublicIPAddressMigrationPhaseCommit
+1. PublicIPAddressMigrationPhase.PublicIPAddressMigrationPhaseCommitted
+1. PublicIPAddressMigrationPhase.PublicIPAddressMigrationPhaseNone
+1. PublicIPAddressMigrationPhase.PublicIPAddressMigrationPhasePrepare
 1. RuleCollectionType.RuleCollectionTypeFirewallPolicyFilterRuleCollection
 1. RuleCollectionType.RuleCollectionTypeFirewallPolicyNatRuleCollection
 1. RuleCollectionType.RuleCollectionTypeFirewallPolicyRuleCollection
@@ -46,6 +58,9 @@
 1. InterfaceIPConfigurationPrivateLinkConnectionProperties.MarshalJSON() ([]byte, error)
 1. ManagedServiceIdentityUserAssignedIdentitiesValue.MarshalJSON() ([]byte, error)
 1. PeerRoute.MarshalJSON() ([]byte, error)
+1. PossibleInterfaceMigrationPhaseValues() []InterfaceMigrationPhase
+1. PossibleInterfaceNicTypeValues() []InterfaceNicType
+1. PossiblePublicIPAddressMigrationPhaseValues() []PublicIPAddressMigrationPhase
 1. ServiceTagInformation.MarshalJSON() ([]byte, error)
 1. ServiceTagInformationPropertiesFormat.MarshalJSON() ([]byte, error)
 1. ServiceTagsListResult.MarshalJSON() ([]byte, error)
@@ -58,3 +73,26 @@
 1. VpnSiteID.MarshalJSON() ([]byte, error)
 1. WatcherPropertiesFormat.MarshalJSON() ([]byte, error)
 1. WebApplicationFirewallPolicyListResult.MarshalJSON() ([]byte, error)
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. AvailablePrivateEndpointType.DisplayName
+1. Delegation.Type
+1. IPAddressAvailabilityResult.IsPlatformReserved
+1. InterfaceIPConfiguration.Type
+1. InterfacePropertiesFormat.MigrationPhase
+1. InterfacePropertiesFormat.NicType
+1. InterfacePropertiesFormat.PrivateLinkService
+1. PublicIPAddressPropertiesFormat.LinkedPublicIPAddress
+1. PublicIPAddressPropertiesFormat.MigrationPhase
+1. PublicIPAddressPropertiesFormat.NatGateway
+1. PublicIPAddressPropertiesFormat.ServicePublicIPAddress
+1. PublicIPPrefixPropertiesFormat.NatGateway
+1. ServiceTagInformationPropertiesFormat.State
+1. Subnet.Type
+1. SubnetPropertiesFormat.ApplicationGatewayIPConfigurations
+1. VirtualNetworkPeering.Type
+1. VirtualNetworkPeeringPropertiesFormat.DoNotVerifyRemoteGateways
+1. VirtualNetworkPeeringPropertiesFormat.ResourceGUID